Reworked to use pytest and entry-points to do away with physical scripts
