+++ .gitignore
@@ -4,7 +4,8 @@ __pycache__/
 *.py[cod]
 *$py.class
 .coverage
-cover/
+.cache
+htmlcov/
 /.eggs
 /build
 /dist
