+++ cgp_seq_input_val/cliutil.py
@@ -13,10 +13,10 @@ def extn_check(parser, choices, fname, readable=False):
         try:
             handle = open(fname, 'r')
             handle.close()
-        except FileNotFoundError as error:  # pragma: no cover
+        except FileNotFoundError as error:
             # can't cover these easily
             parser.error(error)
-    if extn not in choices:  # pragma: no cover
+    if extn not in choices:
         # can't cover these easily
         parser.error("File doesn't end with {}".format(choices))
     return fname
