+++ run_tests.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 set -e
-env/bin/nosetests --with-coverage --cover-erase --cover-html --cover-min-percentage=50 --cover-package=cgp_seq_input_val
+pytest --cov-report term --cov-report html --cov=cgp_seq_input_val --cov-fail-under=50
 set +e
 
 # these should not die:
@@ -8,16 +8,16 @@ set +e
 echo -e "\n##########################"
 echo      "# Running pylint (style) #"
 echo      "##########################"
-env/bin/pylint --output-format=colorized bin/*.py cgp_seq_input_val
+pylint --output-format=colorized cgp_seq_input_val
 
 echo -e "\n#########################################"
 echo      "# Running radon (cyclomatic complexity) #"
 echo      "#########################################"
-env/bin/radon cc -nc bin cgp_seq_input_val
+radon cc -nc cgp_seq_input_val
 
 echo -e "\n#########################################"
 echo      "# Running radon (maintainability index) #"
 echo      "#########################################"
-env/bin/radon mi -s -n B bin cgp_seq_input_val
+radon mi -s -n B cgp_seq_input_val
 
 exit 0 # don't die based on assements of code quality
