+++ setup.py
@@ -11,12 +11,13 @@ config = {
     'author_email': 'cgphelp@sanger.ac.uk',
     'version': '1.1.0',
     'python_requires': '>= 3.3',
-    'setup_requires': ['nose>=1.0'],
+    'setup_requires': ['pytest'],
     'install_requires': ['progressbar2', 'xlrd'],
     'packages': ['cgp_seq_input_val'],
     'package_data': {'cgp_seq_input_val': ['config/*.json']},
-    'scripts': ['bin/normalise_manifest.py', 'bin/validate_manifest.py',
-                'bin/validate_seq_file.py']
+    'entry_points': {
+        'console_scripts': ['cgpSeqInputVal=cgp_seq_input_val.command_line:main'],
+    }
 }
 
 setup(**config)
