+++ tests/test_cgp_seq_input_val_file_meta.py
@@ -1,4 +1,4 @@
-from nose.tools import *
+import pytest
 from cgp_seq_input_val.file_meta import FileValidationError, FileMeta
 import os, sys, tempfile
 
@@ -10,22 +10,22 @@ def teardown():
 
 ### FileMeta tests
 
-@raises(FileValidationError)
 def test_file_meta_file_absent():
-    headers = ["Group_ID", "Sample", "Normal_Tissue", "Group_Control", "Library", "File", "File_2"]
-    details = ["1", "Start", "Y", "Y", "1", "bello.bam"]
-    fm = FileMeta(headers, details, '/')
-    fm.test_files(1)
-
-@raises(FileValidationError)
-def test_file_meta_file_empty():
-    headers = ["Group_ID", "Sample", "Normal_Tissue", "Group_Control", "Library", "File", "File_2"]
-    details = ["1", "Start", "Y", "Y", "1", "bello.bam"]
-    with tempfile.TemporaryDirectory() as tmpd:
-        with open(os.path.join(tmpd, 'bello.bam'), 'w'):
-            pass
-        fm = FileMeta(headers, details, tmpd)
+    with pytest.raises(FileValidationError) as e_info:
+        headers = ["Group_ID", "Sample", "Normal_Tissue", "Group_Control", "Library", "File", "File_2"]
+        details = ["1", "Start", "Y", "Y", "1", "bello.bam"]
+        fm = FileMeta(headers, details, '/')
         fm.test_files(1)
+
+def test_file_meta_file_empty():
+    with pytest.raises(FileValidationError) as e_info:
+        headers = ["Group_ID", "Sample", "Normal_Tissue", "Group_Control", "Library", "File", "File_2"]
+        details = ["1", "Start", "Y", "Y", "1", "bello.bam"]
+        with tempfile.TemporaryDirectory() as tmpd:
+            with open(os.path.join(tmpd, 'bello.bam'), 'w'):
+                pass
+            fm = FileMeta(headers, details, tmpd)
+            fm.test_files(1)
 
 def test_file_meta_get_path():
     headers = ["Group_ID", "Sample", "Normal_Tissue", "Group_Control", "Library", "File", "File_2"]
